[AIP-21] Move sensors.web_hdfs_sensor providers.apache.hdfs.sensors.web_hdfs
