+++ tests/providers/apache/hive/operators/test_hive.py
@@ -30,6 +30,7 @@ from airflow.operators.hive_to_mysql import HiveToMySqlTransfer
 from airflow.operators.hive_to_samba_operator import Hive2SambaOperator
 from airflow.operators.presto_check_operator import PrestoCheckOperator
 from airflow.operators.presto_to_mysql import PrestoToMySqlTransfer
+from airflow.providers.apache.hdfs.sensors.web_hdfs import WebHdfsSensor
 from airflow.providers.apache.hive.operators.hive import HiveOperator
 from airflow.providers.apache.hive.operators.hive_stats import HiveStatsCollectionOperator
 from airflow.providers.apache.hive.sensors.hive_partition import HivePartitionSensor
@@ -37,7 +38,6 @@ from airflow.providers.apache.hive.sensors.metastore_partition import MetastoreP
 from airflow.providers.apache.hive.sensors.named_hive_partition import NamedHivePartitionSensor
 from airflow.sensors.hdfs_sensor import HdfsSensor
 from airflow.sensors.sql_sensor import SqlSensor
-from airflow.sensors.web_hdfs_sensor import WebHdfsSensor
 from airflow.utils import timezone
 
 DEFAULT_DATE = datetime.datetime(2015, 1, 1)
